1552A.cpp
Upload Solutions of any coding Platform

+++ Codeforces Problems/1552A.cpp
@@ -4,7 +4,7 @@
 using namespace std;
 int main()
 {
-long long int t;
+int t;
 cin>>t;
 while(t--)
 {
